# apps/docs/public/openapi.json
{
  "components": {
    "schemas": {
...
      "EconomicActivityProfile": {
        "additionalProperties": false,
        "description": "Actividad económica declarada en el RUC (gActEco, D130-D132).",
        "properties": {
          "codigo": {
            "description": "cActEco (Tabla 3, XSD `tcActEco`).",
            "pattern": "^[0-9A-Z]{1,8}$",
            "title": "Codigo",
            "type": "string"
          },
          "descripcion": {
            "description": "dDesActEco.",
            "maxLength": 300,
            "minLength": 1,
            "title": "Descripcion",
            "type": "string"
          }
        },
        "required": [
          "codigo",
          "descripcion"
        ],
        "title": "EconomicActivityProfile",
        "type": "object"
      },
      "EmitterCreateRequest": {
        "description": "Emitter creation payload.",
        "properties": {
          "csc": {
            "anyOf": [
              {
                "pattern": "^[0-9A-Za-z]{32}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Código de seguridad del contribuyente: 32 caracteres alfanuméricos. Nunca se devuelve.",
            "title": "Csc"
          },
          "csc_id": {
            "anyOf": [
              {
                "pattern": "^[0-9]{1,4}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "IdCSC: de 1 a 4 dígitos entre 1 y 9999; se guarda con cuatro dígitos (`1` → `0001`).",
            "title": "Csc Id"
          },
          "dv": {
            "description": "Dígito verificador del RUC por módulo 11 (D102, 1253).",
            "pattern": "^[0-9]$",
            "title": "Dv",
            "type": "string"
          },
          "external_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id"
          },
          "fiscal_profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EmitterFiscalProfileModel"
              },
              {
                "type": "null"
              }
            ],
            "description": "Perfil fiscal del RUC. Puede cargarse después, pero sin él no se pueden crear documentos."
          },
          "legal_name": {
            "description": "Razón social (D105, XSD `tdNombre`, 4-255).",
            "maxLength": 255,
            "minLength": 4,
            "title": "Legal Name",
            "type": "string"
          },
          "owner_consumer_id": {
            "anyOf": [
              {
                "maxLength": 36,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Consumer Id"
          },
          "ruc": {
            "description": "RUC sin DV (D101, XSD `tRuc`).",
            "maxLength": 8,
            "minLength": 3,
            "pattern": "^[1-9][0-9]*[0-9A-D]?$",
            "title": "Ruc",
            "type": "string"
          },
          "tax_environment": {
            "enum": [
              "test",
              "production"
            ],
            "title": "Tax Environment",
            "type": "string"
          }
        },
        "required": [
          "ruc",
          "dv",
          "legal_name",
          "tax_environment"
        ],
        "title": "EmitterCreateRequest",
        "type": "object"
      },
      "EmitterFiscalProfileModel": {
        "additionalProperties": false,
        "description": "Datos fiscales del emisor tal como figuran en su RUC (gEmis).\n\nEs la única fuente de `gEmis`: sin este perfil no se puede crear un\ndocumento. Se reemplaza completo en cada actualización.",
        "properties": {
          "actividades_economicas": {
            "description": "gActEco (D130), de 1 a 9.",
            "items": {
              "$ref": "#/components/schemas/EconomicActivityProfile"
            },
            "maxItems": 9,
            "minItems": 1,
            "title": "Actividades Economicas",
            "type": "array"
          },
          "domicilio": {
            "$ref": "#/components/schemas/FiscalAddressProfile",
            "description": "Dirección por defecto de los documentos del emisor."
          },
          "establecimientos": {
            "description": "Dirección propia de cada establecimiento que la tenga.",
            "items": {
              "$ref": "#/components/schemas/EstablishmentAddressProfile"
            },
            "maxItems": 999,
            "title": "Establecimientos",
            "type": "array"
          },
          "nombre_fantasia": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 4,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "dNomFanEmi (D106).",
            "title": "Nombre Fantasia"
          },
          "tipo_contribuyente": {
            "description": "iTipCont (D103): 1 persona física, 2 persona jurídica.",
            "enum": [
              1,
              2
            ],
            "title": "Tipo Contribuyente",
            "type": "integer"
          },
          "tipo_regimen": {
            "anyOf": [
              {
                "maximum": 8.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "cTipReg (D104, Tabla 1).",
            "title": "Tipo Regimen"
          }
        },
        "required": [
          "tipo_contribuyente",
          "actividades_economicas",
          "domicilio"
        ],
        "title": "EmitterFiscalProfileModel",
        "type": "object"
      },
...
      "EmitterUpdateRequest": {
        "description": "Emitter update payload.",
        "properties": {
          "csc": {
            "anyOf": [
              {
                "pattern": "^[0-9A-Za-z]{32}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Código de seguridad del contribuyente: 32 caracteres alfanuméricos. Nunca se devuelve.",
            "title": "Csc"
          },
          "csc_id": {
            "anyOf": [
              {
                "pattern": "^[0-9]{1,4}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "IdCSC: de 1 a 4 dígitos entre 1 y 9999; se guarda con cuatro dígitos (`1` → `0001`).",
            "title": "Csc Id"
          },
          "fiscal_profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EmitterFiscalProfileModel"
              },
              {
                "type": "null"
              }
            ],
            "description": "Reemplaza el perfil fiscal completo."
          },
          "legal_name": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 4,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Razón social (D105, XSD `tdNombre`, 4-255).",
            "title": "Legal Name"
          },
          "tax_environment": {
            "anyOf": [
              {
                "enum": [
                  "test",
                  "production"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tax Environment"
          }
        },
        "title": "EmitterUpdateRequest",
        "type": "object"
      },
      "ErrorEnvelope": {
        "description": "Standard error response envelope.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorPayload"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "ErrorPayload": {
        "description": "Standard error payload.",
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "category",
          "correlation_id"
        ],
        "title": "ErrorPayload",
        "type": "object"
      },
...
      "FiscalAddressProfile": {
        "additionalProperties": false,
        "description": "Dirección y contacto del local que emite el DE (D107-D119).",
        "properties": {
          "ciudad": {
            "description": "cCiuEmi (D115).",
            "maximum": 99999.0,
            "minimum": 1.0,
            "title": "Ciudad",
            "type": "integer"
          },
          "complemento_1": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Complemento 1"
          },
          "complemento_2": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Complemento 2"
          },
          "denominacion_sucursal": {
            "anyOf": [
              {
                "maxLength": 30,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "dDenSuc (D119).",
            "title": "Denominacion Sucursal"
          },
          "departamento": {
            "description": "cDepEmi (D111), según `Departamentos_v141.xsd`.",
            "minimum": 1.0,
            "title": "Departamento",
            "type": "integer"
          },
          "descripcion_ciudad": {
            "description": "dDesCiuEmi (D116).",
            "maxLength": 30,
            "minLength": 1,
            "title": "Descripcion Ciudad",
            "type": "string"
          },
          "descripcion_departamento": {
            "anyOf": [
              {
                "maxLength": 16,
                "minLength": 6,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "dDesDepEmi (D112). Opcional: si se envía tiene que coincidir con el código; se guarda siempre la descripción oficial.",
            "title": "Descripcion Departamento"
          },
          "descripcion_distrito": {
            "anyOf": [
              {
                "maxLength": 30,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "dDesDisEmi; obligatoria si hay `distrito`.",
            "title": "Descripcion Distrito"
          },
          "direccion": {
            "description": "dDirEmi (D107).",
            "maxLength": 255,
            "minLength": 1,
            "title": "Direccion",
            "type": "string"
          },
          "distrito": {
            "anyOf": [
              {
                "maximum": 9999.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "cDisEmi.",
            "title": "Distrito"
          },
          "email": {
            "description": "dEmailE.",
            "maxLength": 80,
            "minLength": 3,
            "pattern": "^[0-9a-zA-Z][0-9a-zA-Z.\\-_]*@([0-9a-zA-Z][0-9a-zA-Z\\-_]*\\.)+[a-zA-Z]{2,9}$",
            "title": "Email",
            "type": "string"
          },
          "numero_casa": {
            "description": "dNumCas (D108). `0` solo si la dirección no tiene numeración.",
            "pattern": "^[0-9]{1,6}$",
            "title": "Numero Casa",
            "type": "string"
          },
          "telefono": {
            "description": "dTelEmi (D117), con prefijo.",
            "maxLength": 15,
            "minLength": 6,
            "title": "Telefono",
            "type": "string"
          }
        },
        "required": [
          "direccion",
          "numero_casa",
          "departamento",
          "ciudad",
          "descripcion_ciudad",
          "telefono",
          "email"
        ],
        "title": "FiscalAddressProfile",
        "type": "object"
      },
...
      "SuccessEnvelope": {
        "description": "Standard success response envelope.",
        "properties": {
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          }
        },
        "required": [
          "data",
          "correlation_id"
        ],
        "title": "SuccessEnvelope",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "KilaApiKey": {
        "description": "Private consumer credential. Never expose it to browser clients.",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "API fiscal multi-tenant para emitir, consultar y conciliar documentos electrónicos con SIFEN.",
    "title": "KilaSifen API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/emitters": {
      "post": {
        "operationId": "create_emitter_v1_emitters_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmitterCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing or invalid API key."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The credential lacks the required scope."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found or owned by another tenant."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource conflict or state that does not allow the operation."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request body exceeds the configured size limit; see `details.max_bytes`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed; the request was not processed."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request limit exceeded; retry after `details.retry_after_seconds`.",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying the request.",
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A dependency is temporarily unavailable; retry with backoff."
          }
        },
        "security": [
          {
            "KilaApiKey": []
          }
        ],
        "summary": "Crear emisor",
        "tags": [
          "emitters"
        ]
      }
    },
    "/v1/emitters/{emitter_id}": {
      "get": {
        "operationId": "get_emitter_v1_emitters__emitter_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "emitter_id",
            "required": true,
            "schema": {
              "title": "Emitter Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing or invalid API key."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The credential lacks the required scope."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found or owned by another tenant."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource conflict or state that does not allow the operation."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request body exceeds the configured size limit; see `details.max_bytes`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed; the request was not processed."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request limit exceeded; retry after `details.retry_after_seconds`.",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying the request.",
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A dependency is temporarily unavailable; retry with backoff."
          }
        },
        "security": [
          {
            "KilaApiKey": []
          }
        ],
        "summary": "Obtener emisor",
        "tags": [
          "emitters"
        ]
      },
      "patch": {
        "operationId": "update_emitter_v1_emitters__emitter_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "emitter_id",
            "required": true,
            "schema": {
              "title": "Emitter Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmitterUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing or invalid API key."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The credential lacks the required scope."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found or owned by another tenant."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource conflict or state that does not allow the operation."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request body exceeds the configured size limit; see `details.max_bytes`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed; the request was not processed."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request limit exceeded; retry after `details.retry_after_seconds`.",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying the request.",
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A dependency is temporarily unavailable; retry with backoff."
          }
        },
        "security": [
          {
            "KilaApiKey": []
          }
        ],
        "summary": "Actualizar emisor",
        "tags": [
          "emitters"
        ]
      }
    },
...
    "/v1/emitters/{emitter_id}/deactivate": {
      "post": {
        "operationId": "deactivate_emitter_v1_emitters__emitter_id__deactivate_post",
        "parameters": [
          {
            "in": "path",
            "name": "emitter_id",
            "required": true,
            "schema": {
              "title": "Emitter Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing or invalid API key."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The credential lacks the required scope."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found or owned by another tenant."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource conflict or state that does not allow the operation."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request body exceeds the configured size limit; see `details.max_bytes`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed; the request was not processed."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request limit exceeded; retry after `details.retry_after_seconds`.",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying the request.",
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A dependency is temporarily unavailable; retry with backoff."
          }
        },
        "security": [
          {
            "KilaApiKey": []
          }
        ],
        "summary": "Desactivar emisor",
        "tags": [
          "emitters"
        ]
      }
    },
...
  },
...
}